Physically moved files from analytics dir to core/graph dir
